+++ CHANGELOG.md
@@ -76,3 +76,4 @@
 
 ### 0.7.1
 * make schemaProps in rule outcome optional
+* add publicId to rule tests